requirements.txt
flask
flask_cors
pillow
tensorflow
numpy
gunicorn